Improve test script

+++ make-tests.sh
@@ -1,12 +1,15 @@
 sh build.sh
 
+rm -rf -d terminal
+rm -rf -d sfml
+
 test_template() {
-    ./bin/debug/new $1 test
+    ./bin/debug/new $1 testproj
     echo "\n\n==========================================="
     echo "Testing the $1 template build"
     echo "-----------------------------------"
-    cd $1/test
-    sh scripts/build.sh 
+    cd $1/testproj
+    sh scripts/build.sh
     echo "===========================================\n\n"
     cd ../..
 }
@@ -14,5 +17,8 @@ test_template() {
 test_template terminal
 test_template sfml
 
-rm -rf -d terminal
-rm -rf -d sfml
+if [ "$1" = "delete" ]
+then
+    rm -rf -d terminal
+    rm -rf -d sfml
+fi
